requirements.txt
# === Core Framework ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# === LangGraph / LangChain ===
langgraph>=0.2.40
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph-checkpoint-postgres>=2.0.0

# === LLM Router ===
litellm>=1.50.0

# === Vector Store (Qdrant) ===
qdrant-client>=1.12.0
langchain-qdrant>=0.2.0

# === Database ===
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.0

# === Redis ===
redis[hiredis]>=5.2.0

# === Auth ===
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
email-validator>=2.2.0

# === MCP ===
mcp>=1.1.0
langchain-mcp-adapters>=0.0.3

# === Text Splitting ===
langchain-text-splitters>=0.3.0

# === Document Processing ===
pypdf>=5.1.0
unstructured>=0.15.0

# === Utilities ===
httpx>=0.27.0
python-dotenv>=1.0.0
tenacity>=9.0.0
structlog>=24.4.0

# === Testing ===
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
